+++ Src/Private/Get-AbrAppVolLicense.ps1
@@ -29,10 +29,13 @@ function Get-AbrAPPVolLicense {
     process {
         if ($InfoLevel.AppVolumes.License -ge 1) {
             try {
-                $License = Invoke-RestMethod -SkipCertificateCheck -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/license"
+                if ($PSVersionTable.PSEdition -eq 'Core') {
+                    $License = Invoke-RestMethod -SkipCertificateCheck -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/license"
+                } else {$License = Invoke-RestMethod -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/license"}
+
                 if ($License) {
                     $OutObj = @()
-                    section -Style Heading2 "License Information" {
+                    section -Style Heading3 "License Information" {
                         Switch ($License.license.invalid)
                         {
                             'True' {$LicenseInvalid = 'False' }