+++ Src/Private/Get-AbrAppVolManager.ps1
@@ -29,39 +29,43 @@ function Get-AbrAPPVolManager {
     process {
         if ($InfoLevel.AppVolumes.Managers -ge 1) {
             try {
-                $Managers = Invoke-RestMethod -SkipCertificateCheck -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/manager_services"
+                if ($PSVersionTable.PSEdition -eq 'Core') {
+                    $Managers = Invoke-RestMethod -SkipCertificateCheck -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/manager_services"
+                } else {$Managers = Invoke-RestMethod -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/manager_services"}
+
                 if ($Managers) {
                     $OutObj = @()
-                    section -Style Heading2 "Manager Servers" {
-                        foreach($Manager in $Managers.services) {
-                            try {
-                                $inObj = [ordered] @{
-                                    'Name' = $Manager.name
-                                    'Internal Version' = $Manager.internal_version
-                                    'Product Version' = $Manager.product_version
-                                    'Domain Name' = $Manager.domain_name
-                                    'Computer Name' = $Manager.computer_name
-                                    'Computer FQDN' = $Manager.fqdn
-                                    'Registered' = ConvertTo-TextYN $Manager.registered
-                                    'Secure' = ConvertTo-TextYN $Manager.secure
-                                    'Status' = $Manager.status
-                                    'First Seen At' = $Manager.first_seen_at_human
-                                    'Last Seen At' = $Manager.last_seen_at_human
-                                }
-                                $OutObj = [pscustomobject](ConvertTo-HashToYN $inObj)
+                    section -Style Heading3 "Manager Servers" {
+                        foreach($Manager in $Managers.services | Sort-Object -Property Name) {
+                            section -Style Heading4 $Manager.name {
+                                try {
+                                    $inObj = [ordered] @{
+                                        'Product Version' = $Manager.product_version
+                                        'Internal Version' = $Manager.internal_version
+                                        'Domain Name' = $Manager.domain_name
+                                        'Computer Name' = $Manager.computer_name
+                                        'Computer FQDN' = $Manager.fqdn
+                                        'Registered' = ConvertTo-TextYN $Manager.registered
+                                        'Secure' = ConvertTo-TextYN $Manager.secure
+                                        'Status' = $Manager.status
+                                        'First Seen At' = $Manager.first_seen_at_human
+                                        'Last Seen At' = $Manager.last_seen_at_human
+                                    }
+                                    $OutObj = [pscustomobject](ConvertTo-HashToYN $inObj)
 
-                                $TableParams = @{
-                                    Name = "Manager Server - $($Manager.name)"
-                                    List = $true
-                                    ColumnWidths = 50, 50
+                                    $TableParams = @{
+                                        Name = "Manager Server - $($Manager.name)"
+                                        List = $true
+                                        ColumnWidths = 50, 50
+                                    }
+                                    if ($Report.ShowTableCaptions) {
+                                        $TableParams['Caption'] = "- $($TableParams.Name)"
+                                    }
+                                    $OutObj | Table @TableParams
                                 }
-                                if ($Report.ShowTableCaptions) {
-                                    $TableParams['Caption'] = "- $($TableParams.Name)"
+                                catch {
+                                    Write-PscriboMessage -IsWarning $_.Exception.Message
                                 }
-                                $OutObj | Table @TableParams
-                            }
-                            catch {
-                                Write-PscriboMessage -IsWarning $_.Exception.Message
                             }
                         }
                     }