+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # :arrows_clockwise: VMware AppVolumes As Built Report Changelog
 
-## [0.2.0] - 2022-0?-??
+## [0.2.0] - 2022-10-07
 
 ### Added
 
+- Added packages and programs to the AppStacks section
+- Added managed OUs to the Directory section
+- Added Datastore Members table to the Storage Groups section
+- Enhance SSL certificate validation on PSv5.1
+
+### Changed
+
 - Migrate report to new module format
-  - Implement better error handling
-- Improve report layout
+  - Improvement error handling
+- Improvements on report layout
 
 ## [0.1.0] - 2020-07-06